Refactor: Move Turbine test to a separate file
This commit refactors the Flow testing snippets by moving the Turbine test to its own file. To avoid code duplication, the  and  classes were moved to a new  file, so they can be shared between tests.
