+++ kotlin/src/test/java/com/android/example/flow/test/RepositoryTest.kt
@@ -16,16 +16,10 @@
 
 package com.android.example.flow.test
 
-import app.cash.turbine.test
-import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.toList
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.runTest
 import org.junit.Assert.assertEquals
 import org.junit.Test
@@ -91,65 +85,4 @@ class RepositoryTest {
     }
 }
 
-// [START android_snippets_kotlin_flow_test_repository_and_datasource]
-interface DataSource {
-    fun counts(): Flow<Int>
-}
 
-class Repository(private val dataSource: DataSource) {
-    fun scores(): Flow<Int> {
-        return dataSource.counts().map { it * 10 }
-    }
-}
-
-class FakeDataSource : DataSource {
-    private val flow = MutableSharedFlow<Int>()
-    suspend fun emit(value: Int) = flow.emit(value)
-    override fun counts(): Flow<Int> = flow
-}
-// [END android_snippets_kotlin_flow_test_repository_and_datasource]
-
-@OptIn(ExperimentalCoroutinesApi::class)
-class ContinuousCollectionTest {
-    @Test
-    fun continuouslyCollect() = runTest {
-        // [START android_snippets_kotlin_flow_test_continuously_collect]
-        val dataSource = FakeDataSource()
-        val repository = Repository(dataSource)
-
-        val values = mutableListOf<Int>()
-        backgroundScope.launch(UnconfinedTestDispatcher(testScheduler)) {
-            repository.scores().toList(values)
-        }
-
-        dataSource.emit(1)
-        assertEquals(10, values[0]) // Assert on the list contents
-
-        dataSource.emit(2)
-        dataSource.emit(3)
-        assertEquals(30, values[2])
-
-        assertEquals(3, values.size) // Assert the number of items collected
-        // [END android_snippets_kotlin_flow_test_continuously_collect]
-    }
-
-    @Test
-    fun usingTurbine() = runTest {
-        // [START android_snippets_kotlin_flow_test_using_turbine]
-        val dataSource = FakeDataSource()
-        val repository = Repository(dataSource)
-
-        repository.scores().test {
-            // Make calls that will trigger value changes only within test{}
-            dataSource.emit(1)
-            assertEquals(10, awaitItem())
-
-            dataSource.emit(2)
-            awaitItem() // Ignore items if needed, can also use skip(n)
-
-            dataSource.emit(3)
-            assertEquals(30, awaitItem())
-        }
-        // [END android_snippets_kotlin_flow_test_using_turbine]
-    }
-}
