+++ kotlin/src/test/java/com/android/example/flow/test/fakes.kt
@@ -1,0 +1,39 @@
+/*
+ * Copyright (C) 2025 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.android.example.flow.test
+
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableSharedFlow
+import kotlinx.coroutines.flow.map
+
+// [START android_snippets_kotlin_flow_test_repository_and_datasource]
+interface DataSource {
+    fun counts(): Flow<Int>
+}
+
+class Repository(private val dataSource: DataSource) {
+    fun scores(): Flow<Int> {
+        return dataSource.counts().map { it * 10 }
+    }
+}
+
+class FakeDataSource : DataSource {
+    private val flow = MutableSharedFlow<Int>()
+    suspend fun emit(value: Int) = flow.emit(value)
+    override fun counts(): Flow<Int> = flow
+}
+// [END android_snippets_kotlin_flow_test_repository_and_datasource]
